fix(link account): add cancel button to link account flow
Currently, users who enter the link account flow gets "locked" into that
state. If a user wishes to navigate out of the link account flow or
cancel the flow (because they realize this is the wrong account), the
user is unable to with no options to log out or cancel. This change adds
a cancel button to allow the user to escape from the flow.