fix(jest): mock out imageUtils to prevent a race condition
The animation reducer uses the file uploader callback which throws due
to our implementation of Image. The image api uses live browser
callbacks to render out which causes random tests to fail when the
callback is fired. Instead, mock out the imageUtil library
as images are tested in the integration test library.