Add `unlock_token` to Users
In preparation for eventually adding the ability to lock out user
accounts on repeated failed authentication attempts, we first add a
column with an index on the Users table which will allow users to unlock
their accounts by clicking a link in an email we'll send them.

Because we are on MySQL 8, [adding a column is now an instant
operation](https://dev.mysql.com/doc/refman/8.0/en/innodb-online-ddl-operations.html#online-ddl-column-operations-table),
even on a table as large as our Users table. Unfortunately, [adding an
index is
not](https://dev.mysql.com/doc/refman/8.0/en/innodb-online-ddl-operations.html#online-ddl-index-operations-table).

I did some manual testing on a clone of the production database using an
adhoc, and was able to confirm that adding the new column takes less
than a second and adding the index for that column takes about 35
minutes. During that time, the table is still available for reads and
writes; only schema changes are locked. This means we can safely apply
the change in production without needing to use `gh-ost`, but we will
need to execute the query to add the index outside of the context of our
Rails migrations, which have a 30-second limit.