Make Required ENV Values for `rds` Rake Tasks Actually Required
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/59523,
which made it more explicit when we expect the values we attempt to
fetch from ENV to silently default to `nil` when not present versus when
we actually want to insist on the value being set.

In the Rake tasks for cloning database clusters and deleting clones of
those clusters, we have a couple of required environment variables that
it turns out we were not actually adequately enforcing as required.