chore(deps): force minimist to 1.2.8 to mitigate minimist vulnerability
minimist is an indirect dependency to many of our dependencies. However,
those dependencies have decided not to update themselves which leaves a
vulnerable version of minimist in our code base. To mitigate this,
lock minimist to 1.2.8 globally to force these dependencies to use a
non-vulnerable version.

The following commands were executed to accomplish this:

1. `yarn add minimist@^1.2.8` - to add minimist as a direct dependency
2. `yarn dedupe minimist` - to have yarn force all indirect versions of
minimist to 1.2.8
3. `yarn remove minimist` - to remove the direct dependency

The net result is that minimist is pinned to 1.2.8 for all dependencies
globally.