chore(deps): force globule to 1.3.4 to mitigate globule vulnerability
globule is an indirect dependency to grunt-contrib-watch. However,
grunt-contrib-watch appears to be an abandoned dependency. To mitigate
the lodash vulnerability, I forced the globule dependency to be updated.

The following commands were executed to accomplish this:

1. `yarn add globule@^1.3.4` - to add globule as a direct dependency
2. `yarn dedupe globule` - to have yarn force all indirect versions of
minimist to 1.3.4
3. `yarn remove globule` - to remove the direct dependency

The net result is that globule is pinned to 1.3.4 for all dependencies
globally.