fix: do not propagate the copyright flyout event to prevent a double close
After the react 17 upgrade, the copyright flyout's button's click event
was emitted twice resulting in an open and close (net effect is nothing
is opened). This was due to react 17's event propagation change. This
particular test was missed as it was only tested using an eyes visual
comparison (testing that the copyright showed up).