Temporarily Lock Teacher Accounts on Repeated Failed Authentication (#57515)
* Add `unlock_token` to Users

In preparation for eventually adding the ability to lock out user
accounts on repeated failed authentication attempts, we first add a
column with an index on the Users table which will allow users to unlock
their accounts by clicking a link in an email we'll send them.

Because we are on MySQL 8, [adding a column is now an instant
operation](https://dev.mysql.com/doc/refman/8.0/en/innodb-online-ddl-operations.html#online-ddl-column-operations-table),
even on a table as large as our Users table. Unfortunately, [adding an
index is
not](https://dev.mysql.com/doc/refman/8.0/en/innodb-online-ddl-operations.html#online-ddl-index-operations-table).

I did some manual testing on a clone of the production database using an
adhoc, and was able to confirm that adding the new column takes less
than a second and adding the index for that column takes about 35
minutes. During that time, the table is still available for reads and
writes; only schema changes are locked. This means we can safely apply
the change in production without needing to use `gh-ost`, but we will
need to execute the query to add the index outside of the context of our
Rails migrations, which have a 30-second limit.

* exclude new unlock_token from redshift; as an ephemeral value, it's not particularly useful for long-term analytics

* remove the index before removing the column that it's an index for

* reorder the presentation of information in a comment

* Move Devise Customization into Subdirectory

In preparation for some upcoming work-in-progress changes to [enable
automatic lockout for
teachers(https://github.com/code-dot-org/code-dot-org/pull/57515) and
[allow users to log out of all existing
sessions](https://github.com/code-dot-org/code-dot-org/pull/59763), we'd
like to move our existing customization to Devise functionality into a
subdirectory and associated Module so that the upcoming functionality
can be organized into the same place.

* basic implementation of a custom 'devise lockable' configuration that only locks teachers

* tweak devise to more efficiently integrate with our existing logic

* add translations and templates

* add information about the automatic unlock logic to the manual unlock email

* clean up comments

* add tests and an index to the new column

* add annotations for unlock_token index

* reduce time interval before accounts are automatically unlocked

* only attempt to add an index to the new column on the users table in non-production envrionments; prod will have to be handled specially

* give new tests unique nammes so they don't conflict

* add temporary DCDO gate for new lockout feature

* typo fix

* fix broken test

* default dcdo flag to enabled

we don't want to manually enable this in any environment, and if it needs to be disabled in any that probably means we've got some follow-up work to do anyway

* improve comments

* log some metrics to both cloudwatch and statsig for new user account lock and unlock functionality

* add UserType to logged cloudwatch metrics for new user accounts feature