Update Session Store Evictions Alarm
After a couple weeks of regular evictions, we have a bit of data on which to base our actual alarm threshold. Unfortunately, the data we have is also pretty spiky so it's going to be a balancing act.

[Looking at the past two weeks](https://us-east-1.console.aws.amazon.com/cloudwatch/home?region=us-east-1#metricsV2?graph=~(metrics~(~(~'AWS*2fElastiCache~'Evictions~'CacheClusterId~'ausxf3ubdbzebgu-001~'CacheNodeId~'0001)~(~'...~'ausxf3ubdbzebgu-002~'.~'.)~(~'...~'ausxf3ubdbzebgu-003~'.~'.))~region~'us-east-1~title~'Evictions~stat~'Maximum~view~'timeSeries~start~'2024-08-07T22*3a40*3a00.000Z~end~'2024-08-21T22*3a41*3a00.000Z~period~300)) of data, I propose we increase the time window to 5 minutes and the threshold to 10k; I believe that should be sufficient to avoid false positives while still allowing us to be notified promptly if something does go wrong.

I also simplified this metric to only look at the one node in the cluster that actually performs evictions, and removed some unnecessary `DependsOn` directives (according to `cfn-lint`, these alarms' `SessionStoreGroup` dependency is already enforced by the `Ref`s they use).

Tested with `bundle exec rake stack:lint` and `bundle exec rake stack:validate`