fix(lms): do not set the primary contact info upfront
Previously, the LMS sync process set the user's primary contact info
during the instantiation of a user model. Due to this, when saving the
user model, it will attempt to create the primary contact info
association which doesn't yet exist resulting in a circular model that
tries to create the child and parent at the same time. The net result is
the potential creation of a duplicate user. For instance:

1. User -> primary_contact_info (AuthenticationOption): Create one user
2. primary_contact_info -> User: Create a second user

While saving the user, it is possible that the authentication option may
not yet be created yet which allows the second user to be created. In
most circumstances, the second user will silently fail to be created.

The primary_contact_info is set in the AuthenticationOption
`after_create` hook anyway so the setting of the primary contact info
upfront is not necessary.