Implements a proof-of-concept "regional" site.
This allows studio.code.org/farsi/home to display a truncated page. It
does not yet account for the locale forcing and does not yet exhibit
alterations of content (just removal).

It does show how to have a secondary region as well. In this case,
studio.code.org/india also works and could show different things on its
own.