fix(dependabot): limit open pull requests to only the main app
Dependabot has a hard limit on the number of open pull requests it
tracks at a time. Given the number of PRs that were opened in the misc
scripts, it has resulted in the main app from not being able to open
dependabot PRs against it.

Due to this, temporarily disable PRs except for the main app (apps,
dashboard, and pegasus)