Allows for port to be specified in the database host.
This adds the ability to specify a port via typical ':' syntax to the db_endpoint_* configuration fields. This will allow you to more easily spin up secondary databases for secondary development environments. In particular, this enhances the Docker version of things my allowing the port to be specified via `MYSQL_NATIVE_PORT` such that you can spin it up alongside a native database more easily (and defer uninstalling one if you are migrating.)

This requies some small changes in our utility scripts for getting a console-based REPL for the database connections.

This also updates references to the database user in the Docker context to use `root` since that was what was used to create the databases.