Automatically Refresh ActiveModel's Cached Attributes Data (#60639)
* Add Logic to Refresh ActiveModel's Cached Attributes Data

ActiveModel contains internal logic which will cache information about
valid attributes for the model at initialization time. Unfortunately,
that means if a migration makes a change to a column, the content of the
database might no longer align with that data. That shouldn't usually be
an issue since we're careful not to explicitly reference new columns in
code until the migration has already been applied everywhere, but
queries like `Model.last.attributes` will still implicitly reference the new
columns, which can raise `UnknownAttributeError`s when we attempt to do things like
`Model.new(Model.last.attributes)`.

Rather than trying to police developer usage of the `\.attributes`
method, we'd like to make it safe to use. To that end, this PR proposes
we attempt to handle those `UnknownAttributeError`s gracefully. If we
notice that there are some new migrations in the database when this
error gets raised, we refresh the caches and try to apply the attribute
changes again before allowing the error to bubble up further.

* don't try to fix the attribute method shadowing just yet

* add some basic unit tests

* add a less-basic unit test

* misc comments and cleanup

* fetch count of migrations directly from the underlying database, rather than fetching the actual migrations and counting them in ruby