Fix Violations of and Reenable `Rails/Pluck`
Fix applied automatically with `bundle exec rubocop --autocorrect-all --only Rails/Pluck`

This is a pretty straightforward one:

> This cop enforces the use of `pluck` over `map`.
>
> `pluck` can be used instead of `map` to extract a single key from each element in an enumerable. When called on an Active Record relation, it results in a more efficient query that only selects the necessary key.

In addition to the cleaner code, I think part of the value of this Cop is just reminding us that the `pluck` method exists.

Relying on existing tests to confirm that all existing violations are for data structures that actually support the `pluck` method.

- https://docs.rubocop.org/rubocop-rails/cops_rails.html#railspluck
- https://www.rubydoc.info/gems/rubocop-rails/2.9.1/RuboCop/Cop/Rails/Pluck