Fix Violations of `Style/RedendantParentheses` in HAML
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/60779, which reenabled this Cop and tested that all our Rubocop-linted code passed validation, but unfortunately did not take into account that `haml-lint` uses the same config. Because our automated linting in CI and in the pre-commit hook only checks files changes in the PR, we didn't realize this was opening us up to linting errors in HAML until the staging build.

Fortunately, the total number of violations is relatively small; despite being on a version of `haml-lint` that does not support autocorrection, it was straightforward to make the necessary corrections manually (especially with the help of [`surround.vim`](https://github.com/tpope/vim-surround)!)