test: fix unit test for FinishTeacherAccount
value coming from useSchoolInfo use usIp prop to determine country, so with renderDefault and default params, school_country will be 'US'. also school_zip and school_name default to empty string. tested manually and the user creation route considers empty strings and null as the same, so this is a safe expectation