Monkeypatch `state_machine` Gem
We're using a very old version of `eyes_selenium`, which in turn depends
on an outdated `state_machine` gem which lacks support for Ruby >= 3.2.
More-recent versions of `eyes_selenium` use the new `state_machines`
gem (note the plural), which does support current Ruby. Until we're able
to get our `eyes_selenium` upgraded (which we should do promptly), we
add a temporary patch.