Organize the new RedisSessionStore Middleware into Modules
In order to make it more clear to future developers that it currently contains two distinct (but related) customization to the base functionality.

Also added a few explanatory comments. I'm very open to feedback in particular on whether the comments are a useful addition and whether the module names are accurate and readable.