Entirely Remove Activity Writes
A glitch in our dynamic config system revealed that we are still relying
on a Gatekeeper config value in production in order to avoid attempting
to write entries into the Activities table.

We are at this point committed to not making writes to this table any
more, so rather than relying on a flag to prevent us from throwing
`ActiveModel::RangeError`s on every milestone post, I removed the write
logic entirely. I also removed a few other lines of code only needed for
this deprecated functionality, and identified (but did not touch) one
code fragment that could possibly also be cleaned up with a
more-thorough investigation.

- https://github.com/code-dot-org/code-dot-org/pull/18603
- [Slack thread](https://codedotorg.slack.com/archives/C0T0PNTM3/p1738714542110139)