fix(workshops): handle editing legacy timezones
edits to legacy sessions without time_zone were using the user's local timezone to offset the value sent in the api patch request. checking for the existence of the workshop in the component props determines if the user is editing or creating a workshop. if editing, and the existing session doesn't have a timezone, then it is a legacy session and should not be offset. the timezone label was also misleading for legacy sessions saying the times were UTC, when really they're meant to be local to the timezone where the session was created. update the label to default to 'local'