fix(workshops): use first session timezone when updating workshop
when adding new sessions to an existing workshop using the edit form, the timezone was being set to UTC by the back end since no timezone was sent. this checks the first session for an existing timezone, since all workshops have at least one session created by default. I'm also removing validation that was setting a default time_zone of UTC, instead accepting that if a new session is created without a time_zone that they want the legacy functionality that did not support timezones