Add Alias for Renamed Rake Task
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/64647,
which renamed the Rake `ci` task to `infra:ci` for better
discoverability. Unfortunately, I forgot that our persistent managed
servers will still need to be able to invoke the task at its original
name in order for their first build with the new name to pass.

The original change has already landed on the staging and test servers,
and we manually restarted their builds. This PR is intended to make that
not necessary for tomorrow's DTP.