feat(cache): add support for ISR caching
This PR implements Next.js incremental static regeneration (ISR) for
contentful sourced pages. Previously, pages are generated on-demand
without, necessitating a request to Contentful each time a page is
requested. With this new model, pages are on-demand generated once and
subsequently marked with a `Cache-Control` header of five minutes (until
distributed caching is implemented).

Revalidation has also been implemented in this PR. To accomplish
revalidation, a webhook is fired from Contentful when content is
published which will make a request to the new `/api/revalidate` API. If
the secret revalidation token is correct, the page's cache is
invalidated and a new etag is generated which will enable upstream
caches to also revalidate content.

A followup task is to add support for Redis caching as currently, the
on-demand validation will only invalidate the cache for a single
instance leading to inconsistent pages. For this reason, the cache is
set to five minutes so as to not block progress on our dev stack. Once
distributed caching is implemented, we can subsequently increase this
timeout as all instances will be consistent.