feat(marketing): use oidc instead of iam user
This PR swaps out the implementation of using an iam user to deploy the
marketing app to AWS with the Github Actions OIDC provider. A trust
policy was added which is scoped to the specific Github Environment
`marketing-test` (and in the future, `marketing-prod`) to disallow this
role from being used except when deploying to a protected github
environment.