Clean Up "stitch mode" Options in Eyes Steps
We currently offer devs the ability to specify the stitching mode used
for full-page screenshots by letting them specify either one of the real
stitch modes (`css` or `scroll`) or `none` to disable full-page screenshots
entirely.

Right now, the only time we ever specify stitching mode is when what we
actually want to do is to disable full-page screenshots. So, this PR
proposes we remove the option to specify stitch mode in favor of an
option which will simply disabled full-page screenshots.

Follow-up to https://github.com/code-dot-org/code-dot-org/pull/65664