feat(redis): add redis Next.js page caching
This PR adds support for Next.js to pull cache entries for the full
router page cache via redis. For the infrastructure, the Redis cache
cluster is created via AWS Elasticache and managed with a primary and
single read replica. The instance type chosen is `r7g.xlarge` which is
the default recommended instance type for production workloads. However,
I expect this cache to be over-provisioned and we can reduce the
instance size as needed based on the work load context. For example, the
Code.org cluster may make more sense to use `r7g.xlarge` especially with
many clients hitting `/` by default, but for our smaller sites, a small
cache would be a more frugal option.