fix: run apps tests if shared constants change
changes to shared constants in the lib directory could affect apps tests that rely on the constants that are generated for the front end. this happened on 6/24/25 when a pr that archived a workshop course passed the ci test run because only dashboard tests were run. this pr ensures apps tests run whenever a change to files in lib/cdo/shared_constants is detected