feat(OpenTelemetry): add opentelemetry initializer
This PR adds OpenTelemetry instrumentation to dashboard using the rails
instrumentation defaults. For rack, a customization was made to exclude
static asset paths to reduce the noise sent to the collector. In the
default configuration, traces, metrics, and logs are sent to localhost.
Since this is not enabled by default in any environment (except my test
adhoc), the PR should not introduce anything net new to any environment.

In a follow up, I will add support for an OpenTelemetry collector which
will receive the traces produced by the ruby instrumentation. The
collector would then be responsible for sending the traces to our
OpenTelemetry ingestion provider.