feat(datadog): add datadog rum
This PR adds DataDog rum in parallel to New Relic to trial its usage.
This feature is gated behind a DCDO flag. For initialization, production
is additionally gated by a stage flag. Unfortunately, DCDO is
initialized too late in the bootstrapping cycle to be used for the
DataDog initialization stage.