Allow Overriding Local Webdriver Tests
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/65064, which added logic to use a local webdriver for the first run of all UI and Eyes tests, only falling back to SauceLabs for any tests which fail that first run.

This PR adds a commit flag which can be used to selectively override that behavior if a developer has a specific reason to want even passing tests to run in SauceLabs (for example, if they want to share a video of a newly-added UI test running successfully).

[Slack thread](https://codedotorg.slack.com/archives/C08AMQ869QX/p1753933950644269)