Use Consistent Volume Size Across All Prod Instances
We currently have some logic which which allocates twice as much EBS
capacity on our frontend nodes as we do on the AMI builder. The
reasoning being that frontend nodes end up generating logfiles as they
process traffic, which we expect to take up some amount of space.

What we neglected to consider with this design is that running the build
involves spikes in storage usage, particularly as dashboard assets are
regenerated. It turns out we want a bit of extra storage for both
frontend nodes and the AMI builder, albiet for different reasons.

This PR proposes we resolve that need and simplify our logic; let's just
allocate the same size of volume everywhere.