feat(vite): initialize vite app for code studio
This PR initializes the vite_rails gem for Code Studio. The purpose of
this vite application is to ensure that the in-place modularization of
`apps` will also work in a vite-based application so that we can
continue to make incremental improvements on both the current site as
well as the new one. The vite application will contain CI/CD workflows
to continuously ensure that the vite application will build while
modularization occurs.

For the initial test phase, a new route `/app/*` has been
added which will mirror the current path structure for
pages while maintaining a clear separation between the
current code base and the new. For example,
`/teacher_dashboard/home` will be available at
`/app/teacher_dashboard/home` for the equivalent vite version.
Eventually, we will be able to use a middleware to switch between the
two editions.

The code base is located within the Turborepo managed `frontend`
directory in `frontend/apps/studio`.

To start with, a single entrypoint `application.ts` is used which is
meant to use a router library to sub-chunk to other dynamic entrypoints.
To access this entrypoint, visit
`http://localhost-studio.code.org:3000/vite`.

This entrypoint can also be accessed in vite local mode (no dashboard)
in the `frontend/apps/studio` directory via `yarn start`.