Explicitly cast one parameter in a handcrafted test query
Rails 7.0 adds stronger protections for untyped query params (see
https://github.com/rails/rails/pull/42440), which unfortunately breaks
this pattern of querying a JSON field for an integer value.

The quick fix is to explicitly cast the provided value to an integer
within the query. I'm not sure whether this represents best practice or
if an ActiveRecord::Relation::QueryAttribute would be better here.

See https://github.com/rails/rails/issues/44312 for more discussion of
this change