chore(turbo): use turbo watch mode for dev
Currently, when developing in `frontend`, it triggers a build and executes a persistent `dev` server which watches for changes. However, changing any files in dependencies (such as component library, labs, etc.) does not trigger a rebuild except if you ran `dev` with extra filter flags for the component you wished to run dev on. This results in a poor DevEx as the system should automatically determine what packages to rebuild using its dependency tracking system.

To address this, leverage the `turbo watch` feature which will automatically detect when a specific dependency has changed and trigger its build process. This improves DevEx as contributors will only need to run `yarn dev` on studio which will automatically take care of rebuilding as needed.

This PR also performs an upgrade of Turborepo to the latest version and removes Next.js outputs as we no longer use Next.js here.

- Clarified development instructions for the studio application in README.md.
- Updated package.json scripts for the studio app to use `turbo watch dev --filter @code-dot-org/studio`.
- Adjusted build and dev scripts in music package.json for consistency.
- Updated turbo.json outputs to streamline build process.
- Upgraded turbo dependency version in package.json and yarn.lock.