Use Pregenerated Namespace for User UUIDs
Starting in Rails 7.0, Active Support enforces the RFC4122 specification which requires valid UUIDs as namespace identifiers when generating v5 UUIDs. We are currently using Rails' `secret_key_base` as the namespace when generating UUIDs of user IDs:

https://github.com/code-dot-org/code-dot-org/blob/169cb3d575803069ad585e272d6a4814d3f6982c/dashboard/app/models/user.rb#L1421-L1424

This value is not guaranteed to be a UUID in production, and in other environments like test it is in fact [guaranteed **not** to be a UUID](https://github.com/code-dot-org/code-dot-org/blob/169cb3d575803069ad585e272d6a4814d3f6982c/dashboard/app/models/user.rb#L1421-L1424), resulting in errors like

    ERROR["test_a_get_request_to_get_current_returns_signed_in_user_info", "Api::V1::UsersControllerTest", 572.6412735459999]
     test_a_get_request_to_get_current_returns_signed_in_user_info#Api::V1::UsersControllerTest (572.64s)
    Minitest::UnexpectedError:         ArgumentError: Only UUIDs are valid namespace identifiers
                app/models/user.rb:1420:in `uuid'
                app/controllers/api/v1/users_controller.rb:49:in `current'
                app/controllers/application_controller.rb:433:in `with_global_current_user'
                app/controllers/application_controller.rb:206:in `with_locale'
                test/controllers/api/v1/users_controller_test.rb:270:in `block in <class:UsersControllerTest>'
                test/testing/setup_all_and_teardown_all.rb:36:in `run'

This PR proposes we use the existing "Object ID" namespace, for simplicity.

I'm not certain how these UUIDs are being used in our application, and so I don't know whether we expect them to be consistent between deploys or not. This change will result in us generating different UUIDs for our users than they had before, so we should confirm what effect that will have.

- https://github.com/code-dot-org/code-dot-org/pull/57517
- https://www.ietf.org/rfc/rfc4122.txt
- https://github.com/rails/rails/pull/37682
- https://github.com/rails/rails/blob/7-0-stable/activesupport/CHANGELOG.md#rails-700rc1-december-06-2021