Use Consistent Volume Size Across All Prod Instances (again)
Yet another attempt at upsizing the AMI Builder! See https://github.com/code-dot-org/code-dot-org/pull/69320 for the previous one.

The last attempt was successful in the upsizing, but revealed an existing unrelated issue in the stack which forced us to revert the attempt. Now that that issue has been resolved, we're back to try the upsize again!

See https://github.com/code-dot-org/code-dot-org/pull/69138 for a summary of our deployment plan.