Explicitly create mock session for non-integration test which uses sessions
Specifically, the unit tests for the PartialRegistration helper, which
[implicitly access the session store](https://github.com/code-dot-org/code-dot-org/blob/82af3fa9b6295d52b77076c24e8b8c7131487faa/dashboard/app/models/concerns/partial_registration.rb#L41). We mock up a simple Rack request to facilitate the parts of these tests that depend on it, but until now that has not included a mock session store.

Starting in Rails 7.0, [attempts to write into disabled sessions will raise `ActionDispatch::Request::Session::DisabledSessionError` rather than failing silently](https://github.com/rails/rails/pull/42231). So, in preparation for an eventual upgrade to that version, the time has now come for us to add that mock to this test!