continue to use `add` rather than `merge` to combine AuthenticationOption validations into User
`merge` wraps the errors in a `NestedError` in order to preserve access
to the original error object, which means that the generated errors will
not get deduplicated with each other. We actually depend on that very
deduplication to make the validation failures on the
AuthenticationOption's `email` field show up as though they were
failures on the the User's field.

This is some weird behavior that we probably shouldn't be doing in the
first place, but it means that we do specifically want to avoid using
merge for now. Hopefully we can dramatically simplify this logic at some
point in the future.